Added confirmation dialog to deleting patiënt

+++ src/app/general/sidebar.component.ts
@@ -1,5 +1,6 @@
 import { ActivatedRoute, Router } from '@angular/router';
 import { Component, Input, OnInit } from '@angular/core';
+import { ConfirmationService } from 'primeng/api';
 import { Patient } from '../shared/patient.model';
 import { PatientService } from './../services/patient.service';
 
@@ -20,7 +21,8 @@ export class SidebarComponent implements OnInit {
   constructor(
     public route: ActivatedRoute,
     private patientService: PatientService,
-    private router: Router
+    private router: Router,
+    private confirmationService: ConfirmationService
   ) {}
 
   ngOnInit() {
@@ -55,11 +57,17 @@ export class SidebarComponent implements OnInit {
   }
 
   deletePatient(id: number) {
-    this.patientService.deletePatient(id).subscribe(
-      () => {
-        this.router.navigate(['/patients']);
-      },
-      error => console.log(error)
-    );
+    this.confirmationService.confirm({
+      message:
+        'Bent u zeker dat u deze patiënt (en al zijn/haar medicatie) wilt verwijderen?',
+      accept: () => {
+        this.patientService.deletePatient(id).subscribe(
+          () => {
+            this.router.navigate(['/patients']);
+          },
+          error => console.log(error)
+        );
+      }
+    });
   }
 }
